Changed the type representation declarations
